#17: moved list action / reducer / type to list/item as it belongs to item
