+++ src/components/shop/Editor.js
@@ -4,7 +4,7 @@ import React, {Component} from 'react';
 import styles from './Editor.scss';
 
 import {connect} from 'react-redux';
-import {editItem} from '../../redux/actions/shop/list';
+import {editItem} from '../../redux/actions/shop/list/item';
 import {addItem, updateItem} from '../../redux/actions/shop/editor';
 
 const mapStateToProps = state => {
