+++ src/redux/reducers/index.js
@@ -1,5 +1,5 @@
 import { combineReducers } from 'redux';
-import shopList from './shop/list';
+import shopList from './shop/list/item';
 import shopEditor from './shop/editor';
 
 export default combineReducers({shopList, shopEditor});
