+++ src/redux/types/shop/list/item.js
